Added reset, and other stuff idk
