+++ Configuration/CommandInfo.json
@@ -27,6 +27,12 @@
                 0
             ]
         },
-        "hlt": {}
+        "adc": {
+            "Arguments": [
+                0
+            ]
+        },
+        "hlt": {},
+        "stc": {}
     }
 }
