+++ Interpreter.cs
@@ -17,162 +17,6 @@ namespace Interpreter
             System.Runtime.Serialization.StreamingContext context) : base(info, context) { }
     }
 
-    public class ProcessorFlags : INotifyPropertyChanged
-    {
-        private bool _s = false;
-        private bool _z = false;
-        private bool _ac = false;
-        private bool _p = false;
-        private bool _c = false;
-
-
-        public bool S
-        {
-            get => _s;
-            set
-            {
-                _s = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("S"));
-            }
-        }
-
-
-        public bool Z
-        {
-            get => _z;
-            set
-            {
-                _z = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("Z"));
-            }
-        }
-
-
-        public bool Ac
-        {
-            get => _ac;
-            set
-            {
-                _ac = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("Ac"));
-            }
-        }
-
-
-        public bool P
-        {
-            get => _p;
-            set
-            {
-                _p = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("P"));
-            }
-        }
-
-
-        public bool C
-        {
-            get => _c;
-            set
-            {
-                _c = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("C"));
-            }
-        }
-
-
-        public event PropertyChangedEventHandler? PropertyChanged;
-    }
-
-    public class Registers : INotifyPropertyChanged
-    {
-        private byte _a = 0;
-        private byte _b = 0;
-        private byte _c = 0;
-        private byte _d = 0;
-        private byte _e = 0;
-        private byte _h = 0;
-        private byte _l = 0;
-
-        public byte A
-        {
-            get => _a;
-            set
-            {
-                _a = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("A"));
-            }
-        }
-
-
-
-        public byte B
-        {
-            get => _b;
-            set
-            {
-                _b = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("B"));
-            }
-        }
-
-
-        public byte C
-        {
-            get => _c;
-            set
-            {
-                _c = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("C"));
-            }
-        }
-
-
-        public byte D
-        {
-            get => _d;
-            set
-            {
-                _d = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("D"));
-            }
-        }
-
-
-        public byte E
-        {
-            get => _e;
-            set
-            {
-                _e = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("E"));
-            }
-        }
-
-
-        public byte H
-        {
-            get => _h;
-            set
-            {
-                _h = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("H"));
-            }
-        }
-
-
-        public byte L
-        {
-            get => _l;
-            set
-            {
-                _l = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs("L"));
-            }
-        }
-
-        public event PropertyChangedEventHandler? PropertyChanged;
-    }
     public class Interpreter
     {
 
@@ -330,6 +174,13 @@ namespace Interpreter
         private void _onTimerTimeout(object? source, ElapsedEventArgs e)
         {
             Step();
+        }
+
+        public void ResetProcessor()
+        {
+           _registers.Reset();
+           _memory.Reset();
+           _flags.Reset();
         }
 
         public void Run()
