+++ Processor/IProcessorComponent.cs
@@ -1,0 +1,10 @@
+namespace Interpreter
+{
+    /// <summary>
+    /// Basic functions that all processor components must implement in order to work properly
+    /// </summary>
+    public interface IProcessorComponent
+    {
+        public void Reset();
+    }
+}
