--- ProcessorFlagsDisplay.axaml.cs
@@ -1,14 +1,0 @@
-using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace InterpreterDesktop
-{
-    public partial class ProcessorFlagsDisplay : UserControl
-    {
-        public ProcessorFlagsDisplay()
-        {
-            InitializeComponent();
-        }
-    }
-}
